Converted type test to use tstyche which can also check return types.
