+++ packages/turf-buffer/package.json
@@ -56,7 +56,7 @@
     "build": "tsup --config ../../tsup.config.ts",
     "test": "pnpm run /test:.*/",
     "test:tape": "tsx test.ts",
-    "test:types": "tsc --esModuleInterop --module node16 --moduleResolution node16 --noEmit --strict types.ts"
+    "test:types": "tstyche"
   },
   "devDependencies": {
     "@turf/truncate": "workspace:*",
@@ -66,6 +66,7 @@
     "benchmark": "^2.1.4",
     "load-json-file": "^7.0.1",
     "tape": "^5.9.0",
+    "tstyche": "^6.2.0",
     "tsup": "^8.4.0",
     "tsx": "^4.19.4",
     "typescript": "^5.8.3",
